feat: add new UI components for landing page including Features, Hero, How It Works, Mobile App, Retouch, Header, and Footer
- Implemented Features component showcasing AI generation capabilities.
- Created Hero section with a call to action for downloading the app.
- Added How It Works section detailing the image generation process.
- Developed Mobile App section highlighting mobile features and download options.
- Introduced Retouch component for advanced image editing functionalities.
- Built Header and Footer components for site navigation and information.
- Updated utility components for badges, buttons, cards, and scroll reveal effects.
